+++ src/main/java/net/geforcemods/securitycraft/blocks/SecretStandingSignBlock.java
@@ -25,13 +25,16 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 import net.minecraftforge.common.MinecraftForge;
 
 public class SecretStandingSignBlock extends StandingSignBlock {
+	private final float destroyTimeForOwner;
+
 	public SecretStandingSignBlock(BlockBehaviour.Properties properties, WoodType woodType) {
-		super(properties, woodType);
+		super(OwnableBlock.withReinforcedDestroyTime(properties), woodType);
+		destroyTimeForOwner = OwnableBlock.getStoredDestroyTime();
 	}
 
 	@Override
 	public float getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos) {
-		return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
+		return BlockUtils.getDestroyProgress(super::getDestroyProgress, destroyTimeForOwner, state, player, level, pos);
 	}
 
 	@Override