+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedButtonBlock.java
@@ -4,6 +4,7 @@ import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IReinforcedBlock;
 import net.geforcemods.securitycraft.blockentities.AllowlistOnlyBlockEntity;
+import net.geforcemods.securitycraft.blocks.OwnableBlock;
 import net.geforcemods.securitycraft.misc.OwnershipEvent;
 import net.geforcemods.securitycraft.util.BlockUtils;
 import net.minecraft.core.BlockPos;
@@ -29,11 +30,13 @@ import net.minecraftforge.common.MinecraftForge;
 public class ReinforcedButtonBlock extends ButtonBlock implements IReinforcedBlock, EntityBlock {
 	private final Block vanillaBlock;
 	private final boolean isWoodenButton;
+	private final float destroyTimeForOwner;
 
 	public ReinforcedButtonBlock(boolean isWooden, BlockBehaviour.Properties properties, Block vb) {
-		super(isWooden, properties);
+		super(isWooden, OwnableBlock.withReinforcedDestroyTime(properties));
 		isWoodenButton = isWooden;
 		vanillaBlock = vb;
+		destroyTimeForOwner = OwnableBlock.getStoredDestroyTime();
 	}
 
 	@Override
@@ -46,7 +49,7 @@ public class ReinforcedButtonBlock extends ButtonBlock implements IReinforcedBlo
 
 	@Override
 	public float getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos) {
-		return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
+		return BlockUtils.getDestroyProgress(super::getDestroyProgress, destroyTimeForOwner, state, player, level, pos);
 	}
 
 	@Override