+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedMovingPistonBlock.java
@@ -6,7 +6,6 @@ import java.util.List;
 import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.blockentities.ReinforcedPistonMovingBlockEntity;
-import net.geforcemods.securitycraft.util.BlockUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
@@ -31,11 +30,6 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 public class ReinforcedMovingPistonBlock extends MovingPistonBlock {
 	public ReinforcedMovingPistonBlock(BlockBehaviour.Properties properties) {
 		super(properties);
-	}
-
-	@Override
-	public float getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos) {
-		return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
 	}
 
 	@Override