+++ Changelog.md
@@ -12,8 +12,9 @@
 - Fix: Placing a Secure Redstone Interface as a receiver does not correctly update itself and surrounding blocks
 - Fix: Lenses placed inside a Claymore, Security Camera, or Trophy System sometimes disappear when the world is reloaded
 - Fix: Blocks projected by a Projector are reset to their default state when opening the Projector's screen
-- Fix: The owner of the topmost block of the cage spawned by a Cage Trap is not set correctly 
+- Fix: The owner of the topmost block of the cage spawned by a Cage Trap is not set correctly
 - Fix: Nearby chunks can visually disappear from the world when dismounting a Security Camera that's close to the player
+- Fix: Ownable blocks can be mined by drills or similar blocks from certain other mods
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------